+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/RootNode.java
@@ -50,11 +50,14 @@ import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.ContextLocal;
+import com.oracle.truffle.api.ContextThreadLocal;
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.TruffleContext;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.ContextReference;
 import com.oracle.truffle.api.TruffleLanguage.Env;
+import com.oracle.truffle.api.TruffleLanguage.LanguageReference;
 import com.oracle.truffle.api.TruffleLanguage.ParsingRequest;
 import com.oracle.truffle.api.TruffleStackTraceElement;
 import com.oracle.truffle.api.frame.Frame;
@@ -499,6 +502,71 @@ public abstract class RootNode extends ExecutableNode {
     }
 
     /**
+     * Prepares this {@link RootNode} for compilation. This method is guaranteed to be called at
+     * least once before any code paths are executed as compiled code (see
+     * {@link CompilerDirectives#inCompiledCode()}) for a given compilation tier. It may be called
+     * multiple times per compilation and can be invoked concurrently by multiple compiler threads
+     * without synchronization; therefore, it must be thread-safe.
+     *
+     * By default, this method returns <code>true</code> to indicate sufficient profiling for
+     * compilation, but implementations can return <code>false</code> if not. Returning
+     * <code>false</code> for root compilations will defer the compilation to allow for additional
+     * profiling, whereas otherwise returning <code>false</code> during inlining will disable
+     * inlining for this call site. Any exception thrown will fail the compilation for this
+     * {@link RootNode} permanently.
+     *
+     * Compilations are initiated by the runtime for optimization through partial evaluation and
+     * compilation. The timing and threading of compilations are runtime-specific and should not be
+     * relied upon. This method is not called for Truffle runtimes that do not support compilation,
+     * such as the fallback runtime.
+     *
+     * Compilations may be initiated for the following reasons:
+     * <ul>
+     * <li>The {@link CallTarget} for this {@link RootNode} has been invoked frequently enough to
+     * trigger compilation.
+     * <li>A {@link LoopNode} or {@link BytecodeOSRNode} has executed frequently enough to trigger
+     * OSR compilation.
+     * <li>The compiler is exploring inlining during compilation, preparing to inline a
+     * {@link DirectCallNode} or an indirect call with a partially evaluated constant
+     * {@link CallTarget}.
+     * </ul>
+     *
+     * Work performed in this method counts towards compilation time, not interpretation time.
+     * Implementing this method can help offload expensive computations to compiler threads. Cache
+     * any preparation work to avoid repeating it for every compilation and inlining. To prevent
+     * deadlocks when using synchronization, adhere to the following guidelines:
+     * <ul>
+     * <li>Do not execute guest code or invoke {@link CallTarget#call(Object...)} within this
+     * method.
+     * <li>Avoid acquiring locks shared with execution threads, especially across call boundaries or
+     * loop back-edges when OSR compilation is enabled.
+     * <li>Use {@link AtomicReferenceFieldUpdater} CAS-locking for caching non-expensive preparation
+     * work to minimize contention.
+     * <li>Ensure all synchronization primitives are released upon method exit; use finally blocks
+     * for safety.
+     * </ul>
+     *
+     * Note that during the execution of this method, no language context is entered. Therefore, you
+     * cannot use {@link ContextThreadLocal}, {@link ContextLocal}, {@link LanguageReference}, or
+     * {@link ContextReference}.
+     *
+     * @param rootCompilation <code>true</code> if this is a root compilation; <code>false</code> if
+     *            inlining this root.
+     * @param compilationTier the current compilation tier, as per
+     *            {@link FrameInstance#getCompilationTier()}.
+     * @param lastTier <code>true</code> if this is the last compilation tier for which this
+     *            {@link RootNode} is prepared; <code>false</code> otherwise. Useful for performing
+     *            preparation only at the highest tier.
+     * @return <code>true</code> if the method is sufficiently profiled; <code>false</code>
+     *         otherwise.
+     *
+     * @since 24.2
+     */
+    protected boolean prepareForCompilation(boolean rootCompilation, int compilationTier, boolean lastTier) {
+        return true;
+    }
+
+    /**
      * Is this root node to be considered trivial by the runtime.
      *
      * A trivial root node is defined as a root node that:
@@ -587,7 +655,9 @@ public abstract class RootNode extends ExecutableNode {
      * any thread, even threads unknown to the guest language implementation. It is allowed to
      * create new {@link CallTarget call targets} during preparation of the root node or perform
      * modifications to the {@link TruffleLanguage language} of this root node.
+     * <p>
      *
+     * @see #prepareForCompilation(boolean, int, boolean)
      * @since 20.3
      */
     protected ExecutionSignature prepareForAOT() {